requirements.txt
discord.py
requests
coloredlogs
```